requirements.txt
Django==1.7
python-social-auth==0.2.7
-e git://github.com/cheddarfinancial/cheddarpy.git#egg=cheddarpy
